etl/requirements.txt
apache-airflow[postgres]==2.7.0
pandas
sqlalchemy
psycopg2-binary